requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-chroma>=0.2.0
langchain-community>=0.4.0
langchain-classic>=0.1.0
chromadb>=0.6.0
langchain-text-splitters>=0.3.0
python-dotenv>=1.0.0
streamlit>=1.40.0
requests>=2.31.0
rank_bm25>=0.2.2
jieba>=0.42.1
